Slides/rotary_inverted_pendulum/requirements.txt
matplotlib>=3.10.0
Pillow>=12.0.0
python-pptx>=1.0.2
